+++ commands/delete_test.go
@@ -87,3 +87,42 @@ func TestInteractiveDelete(t *testing.T) {
 		t.Fatalf("actualLength = %q, want %q", actualLength, expectedLength)
 	}
 }
+
+func TestDryRunDelete(t *testing.T) {
+	db := CreateDatabaseForTesting(t)
+	UseDatabaseForTesting(t, db)
+	expectedMachine := "mymachine"
+	expectedService := "myservice"
+	expectedUser := "myuser"
+	expectedPassword := "mypassword"
+	var expectedType PasswordType = "plain"
+	err := initDatabase(db)
+	if err != nil {
+		t.Fatalf("initDatabase() = %q, want nil", err)
+	}
+	context := Context{Database: db}
+	_, err = createPassword(&context, expectedMachine, expectedService, expectedUser, expectedPassword, expectedType)
+	if err != nil {
+		t.Fatalf("createPassword() = %q, want nil", err)
+	}
+	os.Args = []string{"", "delete", "-n", "-m", expectedMachine, "-s", expectedService, "-u", expectedUser}
+	inBuf := new(bytes.Buffer)
+	outBuf := new(bytes.Buffer)
+
+	actualRet := Main(inBuf, outBuf)
+
+	expectedRet := 0
+	if actualRet != expectedRet {
+		t.Fatalf("Main() = %q, want %q", actualRet, expectedRet)
+	}
+	results, err := readPasswords(db, searchOptions{})
+	if err != nil {
+		t.Fatalf("readPasswords() err = %q, want nil", err)
+	}
+	actualLength := len(results)
+	// dry run, so not 0
+	expectedLength := 1
+	if actualLength != expectedLength {
+		t.Fatalf("actualLength = %q, want %q", actualLength, expectedLength)
+	}
+}